Fetch and convert data from rems to AquaChem format

Fetch and convert data from rems to AquaChem format

rems_to_aquachem(
  ems_ids,
  date_range = NULL,
  save = TRUE,
  out_folder = "./",
  out_file = NULL,
  interactive = TRUE
)

Arguments

ems_ids
Character vector. Unique EMS ids

date_range
Character vector. Start and end dates (YYYY-MM-DD)

save
Logical. Whether or not to save the data as csv.

out_folder
Character. Where to save data for AquaChem

out_file
Character. What to call data file for AquaChem. Default is aquachem_DATE.csv

interactive
Logical. Whether or not to ask when caching data.

Value

Outputs an Excel and a CSV file in the format to import into AquaChem.

Details

**`out_folder`** This is how you tell the function where to store the aquachem data.

Examples

# Convert one well and save the data in the working directory
rems_to_aquachem(ems_ids = "E289551")

#> Checking for locally stored historical data...

#> It appears that you already have the most up-to date version of the historic ems data.

#> Checking for locally stored recent data...

#> Fetching data from cache...

#> Called from: rems_to_aquachem(ems_ids = "E289551")
#> debug: d <- d %>% dplyr::mutate_all(.funs = as.character) %>% dplyr::bind_rows(units, 
#>     .)
#> debug: cols <- a$aqua_code[a$aqua_code %in% names(d)]
#> debug: d <- dplyr::select(d, tidyselect::all_of(cols), dplyr::everything())
#> debug: if (save) {
#>     out_file <- file.path(out_folder, out_file)
#>     readr::write_csv(d, out_file, "N/A")
#> }
#> debug: out_file <- file.path(out_folder, out_file)
#> debug: readr::write_csv(d, out_file, "N/A")
#> debug: d

#> # A tibble: 2 x 92
#>   StationID SampleID Sample_Date Analysis_Date Project Watertype shortWatertype
#>   <chr>     <chr>    <chr>       <chr>         <chr>   <chr>     <chr>         
#> 1 ""        ""       ""          ""            ""      ""        ""            
#> 2 "395"     "E28955… "2014-08-1…  NA           "GROUN… "Fresh W…  NA           
#> # … with 85 more variables: Comment <chr>, Reference <chr>,
#> #   Quality_control <chr>, Duplicate_ID <chr>, Labcode <chr>, Location <chr>,
#> #   Geology <chr>, Coord_Lat <chr>, Coord_Long <chr>, X <chr>, Y <chr>,
#> #   Elevation <chr>, Well_Depth <chr>, Screen_Top <chr>, Screen_Mid <chr>,
#> #   Screen_Bottom <chr>, Gradient <chr>, Station_Comment <chr>,
#> #   Sample_Depth <chr>, Temp <chr>, pH_lab <chr>, Cond <chr>, Meas_Alk <chr>,
#> #   CO3 <chr>, HCO3 <chr>, Cl <chr>, F <chr>, Meas_Hardness <chr>, NH4 <chr>,
#> #   NO3 <chr>, NO2 <chr>, SO4 <chr>, Ag_diss <chr>, Al_diss <chr>,
#> #   As_diss <chr>, B <chr>, Ba <chr>, Ca <chr>, Cd_diss <chr>, Cr_diss <chr>,
#> #   Cu_diss <chr>, Fe_diss <chr>, K <chr>, Mg <chr>, Mn_diss <chr>,
#> #   Mo_diss <chr>, Na <chr>, Ni_diss <chr>, Pb_diss <chr>, Sb_diss <chr>,
#> #   Se_diss <chr>, Si <chr>, Sr <chr>, Tl_diss <chr>, U_diss <chr>,
#> #   V_diss <chr>, Zn_diss <chr>, Li <chr>,
#> #   `Alkalinity;_Phenolphthalein_(as_CaCO3)` <chr>, Beryllium_Dissolved <chr>,
#> #   Bismuth_Dissolved <chr>, Bromide_Extractable <chr>, Cobalt_Dissolved <chr>,
#> #   `Hydroxide_Alkalinity_(as_CaCO3)` <chr>,
#> #   `Nitrate(NO3)_+_Nitrite(NO2)_Dissolved` <chr>,
#> #   `Nitrogen_(Kjeldahl)_Total_Dissolved` <chr>,
#> #   Nitrogen_Organic_Total_Dissolved <chr>, Nitrogen_Total_Dissolved <chr>,
#> #   Phosphorus_Total_Dissolved <chr>, `Residue:_Filterable_1_0u_(TDS)` <chr>,
#> #   Sulfur_Dissolved <chr>, Tin_Dissolved <chr>, Titanium_Dissolved <chr>,
#> #   Turbidity <chr>, Zirconium_Dissolved <chr>, Ca_meq <chr>, Mg_meq <chr>,
#> #   Na_meq <chr>, Cl_meq <chr>, HCO3_meq <chr>, SO4_meq <chr>, cations <chr>,
#> #   anions <chr>, charge_balance <chr>, K_meq <chr>

# Convert several wells (this will overwrite the previous data!)
rems_to_aquachem(ems_ids = c("1401030", "1401377"))

#> Checking for locally stored historical data...

#> It appears that you already have the most up-to date version of the historic ems data.

#> Checking for locally stored recent data...

#> Fetching data from cache...

#> Called from: rems_to_aquachem(ems_ids = c("1401030", "1401377"))
#> debug: d <- d %>% dplyr::mutate_all(.funs = as.character) %>% dplyr::bind_rows(units, 
#>     .)
#> debug: cols <- a$aqua_code[a$aqua_code %in% names(d)]
#> debug: d <- dplyr::select(d, tidyselect::all_of(cols), dplyr::everything())
#> debug: if (save) {
#>     out_file <- file.path(out_folder, out_file)
#>     readr::write_csv(d, out_file, "N/A")
#> }
#> debug: out_file <- file.path(out_folder, out_file)
#> debug: readr::write_csv(d, out_file, "N/A")
#> debug: d

#> # A tibble: 14 x 145
#>    StationID SampleID Sample_Date Analysis_Date Project Watertype shortWatertype
#>    <chr>     <chr>    <chr>       <chr>         <chr>   <chr>     <chr>         
#>  1 ""        ""       ""          ""            ""      ""        ""            
#>  2 "075"     "140103… "1987-07-0…  NA           "BACKG… "Fresh W…  NA           
#>  3 "075"     "140103… "1991-08-0…  NA           "BACKG… "Fresh W…  NA           
#>  4 "075"     "140103… "1994-06-0…  NA           "BACKG… "Fresh W…  NA           
#>  5 "075"     "140103… "2001-09-0…  NA           "BACKG… "Fresh W…  NA           
#>  6 "075"     "140103… "2009-11-1…  NA           "BACKG… "Fresh W…  NA           
#>  7 "075"     "140103… "2010-08-0…  NA           "BACKG… "Fresh W…  NA           
#>  8 "075"     "140103… "2016-11-0…  NA           "BACKG… "Fresh W…  NA           
#>  9 "075"     "140103… "2018-06-1…  NA           "BACKG… "Ground …  NA           
#> 10 "203"     "140137… "1987-07-0…  NA           "BACKG… "Fresh W…  NA           
#> 11 "203"     "140137… "1989-10-1…  NA           "BACKG… "Fresh W…  NA           
#> 12 "203"     "140137… "1994-03-2…  NA           "BACKG… "Fresh W…  NA           
#> 13 "203"     "140137… "2016-11-0…  NA           "BACKG… "Fresh W…  NA           
#> 14 "203"     "140137… "2020-06-2…  NA           "BACKG… "Ground …  NA           
#> # … with 138 more variables: Comment <chr>, Reference <chr>,
#> #   Quality_control <chr>, Duplicate_ID <chr>, Labcode <chr>, Location <chr>,
#> #   Geology <chr>, Coord_Lat <chr>, Coord_Long <chr>, X <chr>, Y <chr>,
#> #   Elevation <chr>, Well_Depth <chr>, Screen_Top <chr>, Screen_Mid <chr>,
#> #   Screen_Bottom <chr>, Gradient <chr>, Station_Comment <chr>,
#> #   Sample_Depth <chr>, Temp <chr>, pH_lab <chr>, Cond <chr>, Meas_Alk <chr>,
#> #   CO3 <chr>, HCO3 <chr>, Cl <chr>, F <chr>, Meas_Hardness <chr>, NH4 <chr>,
#> #   NO3 <chr>, NO2 <chr>, SO4 <chr>, Ag_diss <chr>, Al_diss <chr>,
#> #   As_diss <chr>, B <chr>, Ba <chr>, Ca <chr>, Cd_diss <chr>, Cr_diss <chr>,
#> #   Cu_diss <chr>, DO <chr>, Fe_diss <chr>, K <chr>, Mg <chr>, Mn_diss <chr>,
#> #   Mo_diss <chr>, Na <chr>, Ni_diss <chr>, Pb_diss <chr>, pH_field <chr>,
#> #   Sb_diss <chr>, Cond_field <chr>, Se_diss <chr>, Si <chr>, Sr <chr>,
#> #   Tl_diss <chr>, U_diss <chr>, V_diss <chr>, Zn_diss <chr>, Cd_tot <chr>,
#> #   Zn_tot <chr>, Fe_tot <chr>, Cu_tot <chr>, Ni_tot <chr>, Al_tot <chr>,
#> #   Tl_tot <chr>, As_tot <chr>, Sb_tot <chr>, V_tot <chr>, Mo_tot <chr>,
#> #   Ag_tot <chr>, Pb_tot <chr>, Li <chr>, Mn_tot <chr>, U_tot <chr>, Br <chr>,
#> #   `Acidity_pH_4_5_(as_CaCO3)` <chr>, `Acidity_pH_8_3_(as_CaCO3)` <chr>,
#> #   `Alkalinity;_Phenolphthalein_(as_CaCO3)` <chr>, Anion_Sum <chr>,
#> #   Barium_Total <chr>, Beryllium_Dissolved <chr>, Beryllium_Total <chr>,
#> #   Bismuth_Dissolved <chr>, Bismuth_Total <chr>, Boron_Total <chr>,
#> #   Calcium_Total <chr>, Cation_Sum <chr>,
#> #   `Cation-Anion_Balance_%=100*(Cations-Anions)/(Cations+Anions)` <chr>,
#> #   Chromium_Total <chr>, Cobalt_Dissolved <chr>, Cobalt_Total <chr>,
#> #   Cyanide_WAD <chr>, `Hardness_Total_(Total)` <chr>,
#> #   `Hydroxide_Alkalinity_(as_CaCO3)` <chr>, Lithium_Total <chr>,
#> #   Magnesium_Total <chr>, `Nitrate(NO3)_+_Nitrite(NO2)_Dissolved` <chr>,
#> #   `Nitrogen_(Kjeldahl)_Total_Dissolved` <chr>, …

# To specify a date range (Year-Month-Day)
rems_to_aquachem(ems_ids = c("1401030", "1401377"),
                 date_range = c("2000-01-01", "2015-01-01"))

#> Checking for locally stored historical data...

#> It appears that you already have the most up-to date version of the historic ems data.

#> Some EMS ID(s) (1401377) are not present in rems data for this date range, skipping these sites...

#> Called from: rems_to_aquachem(ems_ids = c("1401030", "1401377"), date_range = c("2000-01-01", 
#>     "2015-01-01"))
#> debug: d <- d %>% dplyr::mutate_all(.funs = as.character) %>% dplyr::bind_rows(units, 
#>     .)
#> debug: cols <- a$aqua_code[a$aqua_code %in% names(d)]
#> debug: d <- dplyr::select(d, tidyselect::all_of(cols), dplyr::everything())
#> debug: if (save) {
#>     out_file <- file.path(out_folder, out_file)
#>     readr::write_csv(d, out_file, "N/A")
#> }
#> debug: out_file <- file.path(out_folder, out_file)
#> debug: readr::write_csv(d, out_file, "N/A")
#> debug: d

#> # A tibble: 4 x 132
#>   StationID SampleID Sample_Date Analysis_Date Project Watertype shortWatertype
#>   <chr>     <chr>    <chr>       <chr>         <chr>   <chr>     <chr>         
#> 1 ""        ""       ""          ""            ""      ""        ""            
#> 2 "075"     "140103… "2001-09-0…  NA           "BACKG… "Fresh W…  NA           
#> 3 "075"     "140103… "2009-11-1…  NA           "BACKG… "Fresh W…  NA           
#> 4 "075"     "140103… "2010-08-0…  NA           "BACKG… "Fresh W…  NA           
#> # … with 125 more variables: Comment <chr>, Reference <chr>,
#> #   Quality_control <chr>, Duplicate_ID <chr>, Labcode <chr>, Location <chr>,
#> #   Geology <chr>, Coord_Lat <chr>, Coord_Long <chr>, X <chr>, Y <chr>,
#> #   Elevation <chr>, Well_Depth <chr>, Screen_Top <chr>, Screen_Mid <chr>,
#> #   Screen_Bottom <chr>, Gradient <chr>, Station_Comment <chr>,
#> #   Sample_Depth <chr>, Temp <chr>, pH_lab <chr>, Cond <chr>, Meas_Alk <chr>,
#> #   CO3 <chr>, HCO3 <chr>, Cl <chr>, F <chr>, Meas_Hardness <chr>, NH4 <chr>,
#> #   NO3 <chr>, NO2 <chr>, SO4 <chr>, Ag_diss <chr>, Al_diss <chr>,
#> #   As_diss <chr>, B <chr>, Ba <chr>, Ca <chr>, Cd_diss <chr>, Cr_diss <chr>,
#> #   Cu_diss <chr>, Fe_diss <chr>, K <chr>, Mg <chr>, Mn_diss <chr>,
#> #   Mo_diss <chr>, Na <chr>, Ni_diss <chr>, Pb_diss <chr>, Sb_diss <chr>,
#> #   Se_diss <chr>, Si <chr>, Sr <chr>, Tl_diss <chr>, U_diss <chr>,
#> #   V_diss <chr>, Zn_diss <chr>, Cd_tot <chr>, Zn_tot <chr>, Fe_tot <chr>,
#> #   Cu_tot <chr>, Ni_tot <chr>, Al_tot <chr>, Tl_tot <chr>, As_tot <chr>,
#> #   Sb_tot <chr>, V_tot <chr>, Mo_tot <chr>, Ag_tot <chr>, Pb_tot <chr>,
#> #   Li <chr>, Mn_tot <chr>, U_tot <chr>, Br <chr>,
#> #   `Acidity_pH_4_5_(as_CaCO3)` <chr>, `Acidity_pH_8_3_(as_CaCO3)` <chr>,
#> #   `Alkalinity;_Phenolphthalein_(as_CaCO3)` <chr>, Barium_Total <chr>,
#> #   Beryllium_Dissolved <chr>, Beryllium_Total <chr>, Bismuth_Dissolved <chr>,
#> #   Bismuth_Total <chr>, Boron_Total <chr>, Calcium_Total <chr>,
#> #   Chromium_Total <chr>, Cobalt_Dissolved <chr>, Cobalt_Total <chr>,
#> #   Cyanide_WAD <chr>, `Hardness_Total_(Total)` <chr>, Lithium_Total <chr>,
#> #   Magnesium_Total <chr>, `Nitrate(NO3)_+_Nitrite(NO2)_Dissolved` <chr>,
#> #   Nitrogen_Ammonia_Total <chr>, `Nitrogen_Kjel_Tot(N)` <chr>,
#> #   `Nitrogen_Organic-Total` <chr>, Nitrogen_Total <chr>,
#> #   `Phosphorus_Ort_Dis-P` <chr>, Phosphorus_Total <chr>,
#> #   Phosphorus_Total_Dissolved <chr>, Potassium_Total <chr>, …

# To name the ouput (extension should be .csv, .txt, or .dat)
rems_to_aquachem(ems_ids = "1401030", out_file = "water_quality01.csv")

#> Checking for locally stored historical data...

#> It appears that you already have the most up-to date version of the historic ems data.

#> Checking for locally stored recent data...

#> Fetching data from cache...

#> Called from: rems_to_aquachem(ems_ids = "1401030", out_file = "water_quality01.csv")
#> debug: d <- d %>% dplyr::mutate_all(.funs = as.character) %>% dplyr::bind_rows(units, 
#>     .)
#> debug: cols <- a$aqua_code[a$aqua_code %in% names(d)]
#> debug: d <- dplyr::select(d, tidyselect::all_of(cols), dplyr::everything())
#> debug: if (save) {
#>     out_file <- file.path(out_folder, out_file)
#>     readr::write_csv(d, out_file, "N/A")
#> }
#> debug: out_file <- file.path(out_folder, out_file)
#> debug: readr::write_csv(d, out_file, "N/A")
#> debug: d

#> # A tibble: 9 x 144
#>   StationID SampleID Sample_Date Analysis_Date Project Watertype shortWatertype
#>   <chr>     <chr>    <chr>       <chr>         <chr>   <chr>     <chr>         
#> 1 ""        ""       ""          ""            ""      ""        ""            
#> 2 "075"     "140103… "1987-07-0…  NA           "BACKG… "Fresh W…  NA           
#> 3 "075"     "140103… "1991-08-0…  NA           "BACKG… "Fresh W…  NA           
#> 4 "075"     "140103… "1994-06-0…  NA           "BACKG… "Fresh W…  NA           
#> 5 "075"     "140103… "2001-09-0…  NA           "BACKG… "Fresh W…  NA           
#> 6 "075"     "140103… "2009-11-1…  NA           "BACKG… "Fresh W…  NA           
#> 7 "075"     "140103… "2010-08-0…  NA           "BACKG… "Fresh W…  NA           
#> 8 "075"     "140103… "2016-11-0…  NA           "BACKG… "Fresh W…  NA           
#> 9 "075"     "140103… "2018-06-1…  NA           "BACKG… "Ground …  NA           
#> # … with 137 more variables: Comment <chr>, Reference <chr>,
#> #   Quality_control <chr>, Duplicate_ID <chr>, Labcode <chr>, Location <chr>,
#> #   Geology <chr>, Coord_Lat <chr>, Coord_Long <chr>, X <chr>, Y <chr>,
#> #   Elevation <chr>, Well_Depth <chr>, Screen_Top <chr>, Screen_Mid <chr>,
#> #   Screen_Bottom <chr>, Gradient <chr>, Station_Comment <chr>,
#> #   Sample_Depth <chr>, Temp <chr>, pH_lab <chr>, Cond <chr>, Meas_Alk <chr>,
#> #   CO3 <chr>, HCO3 <chr>, Cl <chr>, F <chr>, Meas_Hardness <chr>, NH4 <chr>,
#> #   NO3 <chr>, NO2 <chr>, SO4 <chr>, Ag_diss <chr>, Al_diss <chr>,
#> #   As_diss <chr>, B <chr>, Ba <chr>, Ca <chr>, Cd_diss <chr>, Cr_diss <chr>,
#> #   Cu_diss <chr>, DO <chr>, Fe_diss <chr>, K <chr>, Mg <chr>, Mn_diss <chr>,
#> #   Mo_diss <chr>, Na <chr>, Ni_diss <chr>, Pb_diss <chr>, pH_field <chr>,
#> #   Sb_diss <chr>, Cond_field <chr>, Se_diss <chr>, Si <chr>, Sr <chr>,
#> #   Tl_diss <chr>, U_diss <chr>, V_diss <chr>, Zn_diss <chr>, Cd_tot <chr>,
#> #   Zn_tot <chr>, Fe_tot <chr>, Cu_tot <chr>, Ni_tot <chr>, Al_tot <chr>,
#> #   Tl_tot <chr>, As_tot <chr>, Sb_tot <chr>, V_tot <chr>, Mo_tot <chr>,
#> #   Ag_tot <chr>, Pb_tot <chr>, Li <chr>, Mn_tot <chr>, U_tot <chr>, Br <chr>,
#> #   `Acidity_pH_4_5_(as_CaCO3)` <chr>, `Acidity_pH_8_3_(as_CaCO3)` <chr>,
#> #   `Alkalinity;_Phenolphthalein_(as_CaCO3)` <chr>, Anion_Sum <chr>,
#> #   Barium_Total <chr>, Beryllium_Dissolved <chr>, Beryllium_Total <chr>,
#> #   Bismuth_Dissolved <chr>, Bismuth_Total <chr>, Boron_Total <chr>,
#> #   Calcium_Total <chr>, Cation_Sum <chr>,
#> #   `Cation-Anion_Balance_%=100*(Cations-Anions)/(Cations+Anions)` <chr>,
#> #   Chromium_Total <chr>, Cobalt_Dissolved <chr>, Cobalt_Total <chr>,
#> #   Cyanide_WAD <chr>, `Hardness_Total_(Total)` <chr>,
#> #   `Hydroxide_Alkalinity_(as_CaCO3)` <chr>, Lithium_Total <chr>,
#> #   Magnesium_Total <chr>, `Nitrate(NO3)_+_Nitrite(NO2)_Dissolved` <chr>,
#> #   `Nitrogen_(Kjeldahl)_Total_Dissolved` <chr>, …

# To save the output to a specific folder
# (here, the Outputs folder in the Rcode folder on the H drive)
# Note that the out_folder must exist or you'll get an error
if (FALSE) rems_to_aquachem(ems_ids = c("1401030", "1401377"),
                 out_folder = "H:\\Rcode\\Outputs/")

# All together now!
if (FALSE) rems_to_aquachem(ems_ids = c("1401030", "1401377"),
                 date_range = c("2000-01-01", "2015-01-01"),
                 out_file = "water_quality05.csv",
                 out_folder = "H:\\Rcode\\Outputs/")

# Clean up
unlink("water_quality01.csv")
unlink(paste0("aquachem_", Sys.Date(), ".csv"))